+++ scalajsReactInterop/src/main/scala/slinky/scalajsreact/Converters.scala
@@ -1,7 +1,7 @@
 package slinky.scalajsreact
 
 import japgolly.scalajs.react.component.Generic.UnmountedRaw
-import japgolly.scalajs.react.raw.ReactNode
+import japgolly.scalajs.react.raw.React.Element
 import japgolly.scalajs.react.vdom.{TagOf, VdomNode}
 import japgolly.scalajs.react.vdom.html_<^._
 
@@ -23,6 +23,6 @@ object Converters {
   }
 
   implicit def componentInstanceToVdom[T](component: T)(implicit ev: T => ReactElement): VdomNode = {
-    VdomNode(ev(component).asInstanceOf[ReactNode])
+    VdomNode(ev(component).asInstanceOf[Element])
   }
 }